dom/grid/test/chrome/test_grid_track_state.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
<style>
body {
	margin: 40px;
}
.wrapper {
	display: grid;
	width: 600px;
	grid-gap: 0px;
	background-color: #f00;
}
.grid1 {
	grid-template-columns: 50px 0 repeat(auto-fit, 100px);
}
.grid2 {
	grid-template-columns: 50px 0 [real-before] repeat(auto-fit, [before] 100px [after]) [real-after];
}
.grid3 {
	grid-template-columns: repeat(3, 66px) [real-before] repeat(auto-fit, [before] 100px [after]) [real-after];
}
.grid4 {
	grid-template-columns: repeat(2, 100px) repeat(auto-fill, 50px);
}
.grid5 {
	grid-template-columns: [real-before] repeat(auto-fit, [before] 100px [after]) [real-after];
}
.box {
	background-color: #444;
	color: #fff;
}
.a {
	grid-column: auto;
}
.b {
	grid-column: 4;
}
.c {
	grid-column: 6;
}
.d {
	grid-column: 7;
}

</style>

<script>
'use strict';

SimpleTest.waitForExplicitFinish();

function runTests() {
	let wrapper = document.getElementById("wrapper1");
	let grid = wrapper.getGridFragments()[0];
	
	// test auto-fit count after removal
	is(grid.cols.tracks.length, 3, "Grid column track array compensates for removed auto-fit columns.");
	
	// test resolved value of grid-template-columns
	let templateColumnsText = getComputedStyle(wrapper).gridTemplateColumns;
	is(templateColumnsText, "50px 0px 0px 100px 0px 0px 0px",
		"Resolved value of grid-template-columns reports removed auto-fits as '0px'.");
	
	is(grid.cols.tracks[0].breadth, 50, "Column 1 is breadth 50.");
	is(grid.cols.tracks[1].breadth, 0, "Column 2 is breadth 0.");
	is(grid.cols.tracks[2].breadth, 100, "Column 3 is breadth 100.");

	
	wrapper = document.getElementById("wrapper2");
	grid = wrapper.getGridFragments()[0];
	
	// test resolved value of grid-template-columns
	templateColumnsText = getComputedStyle(wrapper).gridTemplateColumns;
	is(templateColumnsText, "50px 0px [real-before before] 0px [after before] 100px [after before] 0px [after before] 100px [after before] 0px [after real-after]",
		"Resolved value of grid-template-columns reports lines for removed tracks.");
	
	
	wrapper = document.getElementById("wrapper3");
	grid = wrapper.getGridFragments()[0];
	
	// test resolved value of grid-template-columns
	templateColumnsText = getComputedStyle(wrapper).gridTemplateColumns;
	is(templateColumnsText, "66px 66px 66px [real-before before] 100px [after before] 0px [after before] 100px [after before] 100px [after real-after]",
		"Resolved value of grid-template-columns reports lines for removed tracks.");
	
	
	wrapper = document.getElementById("wrapper4");
	grid = wrapper.getGridFragments()[0];
	
	// test auto-fill count of tracks
	is(grid.cols.tracks.length, 10, "Grid column track array respects auto-fill columns.");
	
	if (grid.cols.tracks.length == 10) {
		// test for static and repeat
		is(grid.cols.tracks[1].state, "static", "Grid column track 2 is marked as static.");
		is(grid.cols.tracks[2].state, "repeat", "Grid column track 3 is marked as repeat.");
	}
	
	
	wrapper = document.getElementById("wrapper5");
	grid = wrapper.getGridFragments()[0];
	
	// test resolved value of grid-template-columns
	templateColumnsText = getComputedStyle(wrapper).gridTemplateColumns;
	is(templateColumnsText, "[real-before before] 0px [after before] 0px [after before] 0px [after before] 0px [after before] 0px [after before] 0px [after real-after]",
		"Resolved value of grid-template-columns no longer lists 'none' when all auto-fit tracks are empty.");
	
	SimpleTest.finish();
}
</script>
</head>
<body onLoad="runTests();">

	<div id="wrapper1" class="wrapper grid1">
		<div id="boxB" class="box b">B</div>
	</div>
		
	<br/>
	
	<div id="wrapper2" class="wrapper grid2">
		<div id="boxB" class="box b">B</div>
		<div id="boxC" class="box c">C</div>
	</div>
		
	<br/>
	
	<div id="wrapper3" class="wrapper grid3">
		<div id="boxB" class="box b">B</div>
		<div id="boxC" class="box c">C</div>
		<div id="boxD" class="box d">D</div>
	</div>
		
	<br/>
	
	<div id="wrapper4" class="wrapper grid4">
		<div id="boxA" class="box a">A</div>
	</div>
	
	<br/>
	
	<div id="wrapper5" class="wrapper grid5">
	</div>
	
</body>
</html>
